refactor(cli): migrate earthdata tooling onto discovery
Move earthdata's CMR refresher/writer/curated_ids/prober/deep_prober/emitter to
earthlens.earthdata.cli (including the format/output_kind heuristics that were in
core's stanza module). Relocate the CLI tests to earthlens-imagery. Shrink the
guard allow-list (earthdata).
